Skip simulation when number of mutations was 0; just inherit the parent fitness.
Change a lot of the --options to make naming more consistent.
                                                                                                                                                                                                      Remove random.setstate stuff - we can't reproduce cluster evolved stuff anyway,
and there seems to be some problem at the moment with the steady state
evolution reproducing the exact same child bpgs and thus not exploring the
solution space properly.

Remove temporary .dat files after plotting.

Fix ranges on some of the plots.